Add --no-tls option.

+++ start-trow.sh
@@ -2,7 +2,7 @@
 
 # Temporary launch script to bootstrap certs that should go away
 
-if [[ ! -f ./certs/domain.key && ! -f ./certs/ca.crt ]]; then
+if [[ $@ != *"--no-tls"* && ! -f ./certs/domain.key && ! -f ./certs/ca.crt ]]; then
   echo "No certs found, creating new ones"
   mkdir ./certs || true
   cp install/self-cert/* ./certs/
@@ -10,7 +10,7 @@ if [[ ! -f ./certs/domain.key && ! -f ./certs/ca.crt ]]; then
   ./make-certs.sh
   cd /
 fi
-echo "Running $@"
+echo "Running /trow $@"
 
 exec /trow "$@"
